Fix bytecode complication on llvm-enabled PG 11+.
For each files from the extension pgxs infrastructure attempts to
build a bytecode from the corresponding C source. It doesn't always
work, since the rules in Makefile.global discard all CFLAGS, therefore
in our case we need to passt the type name via the CPPFLAGS that are
honoted by the install_llvm_module.